Make video detail retrievers more modular
